# Performance optimization for AnalyzeSolutionGenerator