## Description
* WorkspaceBuilderHelper.BuildIncremental - if a project reference was already built and added to a workspace, use the existing ProjectAnalysisResult instead of building again.

## Supplemental testing

### BenchmarkDotNet results for CodeAnalyzerByLanguage.AnalyzeSolutionGenerator method execution
``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.20348.887)
Intel Xeon Platinum 8375C CPU 2.90GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.400
  [Host]     : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT AVX2
  Job-WHKMPK : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT AVX2

LaunchCount=1  RunStrategy=Monitoring  

```
|                   Method |     Mean |   Error |  StdDev | Ratio | RatioSD |        Gen0 |       Gen1 |      Gen2 | Allocated | Alloc Ratio |
|------------------------- |---------:|--------:|--------:|------:|--------:|------------:|-----------:|----------:|----------:|------------:|
| AnalyzeSolutionGenerator (Before changes) | 141.83 s | 3.001 s | 1.985 s |  3.55 |    0.05 | 202000.0000 | 89000.0000 | 1000.0000 |   5.73 GB |        1.17 |
| AnalyzeSolutionGenerator (After changes) | 62.38 s | 0.619 s | 0.409 s |  1.59 |    0.02 | 169000.0000 | 77000.0000 | 1000.0000 |   4.85 GB |        0.99 |

## Additional context
### Benchmarking code (in my fork; not in this PR) - 
1.  [Before changes](https://github.com/neerajhanda/codelyzer/tree/benchmark-main/tst/Codelyzer.Analysis.Benchmarks)
2. [After changes](https://github.com/neerajhanda/codelyzer/tree/benchmark-performance-enhancements/tst/Codelyzer.Analysis.Benchmarks)

### Unit tests - 
10 unit tests were failing on the main branch before my changes. No additional tests are failing after my changes.

---
*By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.*